Client : implemented TAB key (via +button8) as "free camera" mode, holding it down allows player to look around with the mouse (turning off the cursor).
A minor annoyance is that when releasing TAB, the pointer returns to the
wrong position (the center of the screen).

A _MAJOR_ annoyance is that binding this to a mouse button does not work
properly, pressing it once works, but next button down event does nothing,
then next one works, etc....  Seems the problem is in SDL library, as the
SDL_MOUSEBUTTONDOWN event is not received.